requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
python-dotenv==1.0.1
androguard==4.1.2
